chore: update logging to json based logging
