+++ apiserver/plane/celery.py
@@ -1,7 +1,15 @@
+# Python imports
 import os
+import logging
+
+# Third party imports
 from celery import Celery
-from plane.settings.redis import redis_instance
+from pythonjsonlogger.jsonlogger import JsonFormatter
+from celery.signals import after_setup_logger, after_setup_task_logger
 from celery.schedules import crontab
+
+# Module imports
+from plane.settings.redis import redis_instance
 
 # Set the default Django settings module for the 'celery' program.
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "plane.settings.production")
@@ -46,6 +54,28 @@ app.conf.beat_schedule = {
         "schedule": crontab(hour=2, minute=30),  # UTC 02:30
     },
 }
+
+
+# Setup logging
+@after_setup_logger.connect
+def setup_loggers(logger, *args, **kwargs):
+    formatter = JsonFormatter(
+        '"%(levelname)s %(asctime)s %(module)s %(name)s %(message)s'
+    )
+    handler = logging.StreamHandler()
+    handler.setFormatter(fmt=formatter)
+    logger.addHandler(handler)
+
+
+@after_setup_task_logger.connect
+def setup_task_loggers(logger, *args, **kwargs):
+    formatter = JsonFormatter(
+        '"%(levelname)s %(asctime)s %(module)s %(name)s %(message)s'
+    )
+    handler = logging.StreamHandler()
+    handler.setFormatter(fmt=formatter)
+    logger.addHandler(handler)
+
 
 # Load task modules from all registered Django app configs.
 app.autodiscover_tasks()
