+++ apiserver/plane/middleware/logger.py
@@ -10,8 +10,10 @@ from rest_framework.request import Request
 
 # Module imports
 from plane.utils.ip_address import get_client_ip
+from plane.db.models import APIActivityLog
 
-api_logger = logging.getLogger("plane.api")
+
+api_logger = logging.getLogger("plane.api.request")
 
 
 class RequestLoggerMiddleware:
@@ -69,3 +71,41 @@ class RequestLoggerMiddleware:
 
         # return the response
         return response
+
+
+class APITokenLogMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        request_body = request.body
+        response = self.get_response(request)
+        self.process_request(request, response, request_body)
+        return response
+
+    def process_request(self, request, response, request_body):
+        api_key_header = "X-Api-Key"
+        api_key = request.headers.get(api_key_header)
+        # If the API key is present, log the request
+        if api_key:
+            try:
+                APIActivityLog.objects.create(
+                    token_identifier=api_key,
+                    path=request.path,
+                    method=request.method,
+                    query_params=request.META.get("QUERY_STRING", ""),
+                    headers=str(request.headers),
+                    body=(request_body.decode("utf-8") if request_body else None),
+                    response_body=(
+                        response.content.decode("utf-8") if response.content else None
+                    ),
+                    response_code=response.status_code,
+                    ip_address=get_client_ip(request=request),
+                    user_agent=request.META.get("HTTP_USER_AGENT", None),
+                )
+
+            except Exception as e:
+                api_logger.exception(e)
+                # If the token does not exist, you can decide whether to log this as an invalid attempt
+
+        return None
